Remove sensitive data from client_secret.json
I can't exactly propose an edit for this, but I recommend not having the private keys stored in a public file. I found this repo from twitter, I'm guessing others might be able to as well. 
Here's an alternative (but I'm sure there are other ways to store private access keys that work as well): https://help.github.com/en/actions/configuring-and-managing-workflows/creating-and-storing-encrypted-secrets#creating-encrypted-secrets-for-a-repository